backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.30,<1.0
supabase>=2.0,<3.0
python-dotenv>=1.0,<2.0
pydantic>=2.0,<3.0
openai>=1.0,<3.0
